add a fuzzing test
