+++ Cargo.toml
@@ -18,6 +18,7 @@ members = [
     "models/rwuerstchen",
     "models/segment-anything-rs",
     "models/kalosm-ocr",
+    "models/tokenator",
     "interfaces/kalosm",
     "interfaces/kalosm-language",
     "interfaces/language-model",
