--- models/tokenizer/Cargo.toml
@@ -1,31 +1,0 @@
-[package]
-name = "fast-bpe"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-serde = { version = "1.0.205", features = ["derive"] }
-serde_json = "1.0.122"
-trie-rs = "0.4.2"
-bumpalo = { version = "3.16.0", features = ["collections", "allocator_api", "allocator-api2"] }
-hashbrown = { version = "0.14.5", features = ["nightly"] }
-rustc-hash = "2.0.0"
-colored = "2.1.0"
-serde-big-array = "0.5.1"
-postcard = { version = "1.0.10", features = ["use-std"] }
-regex = { version = "1.10.6", features = ["perf-dfa-full"] }
-reqwest = { version = "0.12.7", features = ["blocking"] }
-tracing = "0.1.40"
-
-[profile.release]
-lto = true
-codegen-units = 1
-# debug = true
-
-[dev-dependencies]
-criterion = "0.3"
-tokenizers = "0.20.0"
-
-[[bench]]
-name = "tokenize"
-harness = false
